requirements.txt
asgiref==3.5.2
Authlib==1.0.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.5
Django==4.0.5
django-allauth==0.51.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
ffmpeg-python==0.2.0
filelock==3.7.1
future==0.18.2
gcloud==0.18.3
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
jwcrypto==1.3.1
moviepy==1.0.3
mysql-connector-python==8.0.29
nodeenv==1.7.0
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.0
Pillow==9.2.0
platformdirs==2.5.2
proglog==0.1.10
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==3.3.2
python3-openid==3.2.0
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tqdm==4.64.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.16.2
wrapt==1.14.1
